gl: fix incorrect extension PFN ranges for GL/GLES2 split extensions
GL_KHR_debug has different PFN ranges depending on if you're running on
desktop GL or GLES, and their names collide with the ES3.2 versions.

For example, glDebugMessageControl:

- Defined as 'glDebugMessageControl' for GL 4.3
- Defined as 'glDebugMessageControl' for ES 3.2
- Defined as 'glDebugMessageControl' for GL_KHR_debug on GL
- Defined as 'glDebugMessageControlKHR' for GL_KHR_debug on GLES

This ambiguity causes us to incorrectly resolve the right entry point on
a GLES 3.1 context -- ANGLE has API support for ES 3.2, but those entry
points do not work on an ES 3.1 context. So we do not want to load the
non-KHR suffixed versions unless we're on an ES 3.2 context.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>