c: replace standard libc 'qsort' with our own uint64_t sort function
The main problem with libc's qsort is that it has an un-inlinable
comparator function that always boils to an indirect function call. No
matter the underlying sorting algorithm, you will always pay the penalty
of such a call.

Writing our own sort in here helps because we know the exact size of the
data type and how to compare it *within* the sort loop bodies. And we
can use a reasonably optimal sorting algorithm for a 0-500 element
range, as well.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>